Move GTM and Sentry to App.js
